+++ packages/google_maps_flutter/google_maps_flutter/example/test_driver/google_map_inspector.dart
@@ -37,13 +37,6 @@ class GoogleMapInspector {
     return zoomLevel;
   }
 
-  Future<double> getDistance(double meters, LatLng latLng) async {
-    final double distance = await _channel.invokeMethod<double>(
-        'map#getDistance',
-        <String, dynamic>{"meter": meters, "location": latLng.toJson()});
-    return distance;
-  }
-
   Future<bool> isZoomGesturesEnabled() async {
     return await _channel.invokeMethod<bool>('map#isZoomGesturesEnabled');
   }