+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.2
 
-* Added getDistance api
+* Added getPoint projections api for iOS
 
 ## 1.0.1